Move TrackGeometryDataset into track_dataset module
