+++ OTAnalytics/plugin_datastore/track_store.py
@@ -22,14 +22,13 @@ from OTAnalytics.domain.geometry import (
     calculate_direction_vector,
 )
 from OTAnalytics.domain.section import Section, SectionId
-from OTAnalytics.domain.track import (
+from OTAnalytics.domain.track import Detection, Track, TrackId
+from OTAnalytics.domain.track_dataset import (
     TRACK_GEOMETRY_FACTORY,
-    Detection,
-    Track,
+    IntersectionPoint,
+    TrackDataset,
     TrackGeometryDataset,
-    TrackId,
 )
-from OTAnalytics.domain.track_dataset import IntersectionPoint, TrackDataset
 from OTAnalytics.domain.types import EventType
