+++ tests/OTAnalytics/plugin_datastore/conftest.py
@@ -3,7 +3,11 @@ from unittest.mock import Mock
 
 import pytest
 
-from OTAnalytics.domain.track import TRACK_GEOMETRY_FACTORY, Track, TrackGeometryDataset
+from OTAnalytics.domain.track import Track
+from OTAnalytics.domain.track_dataset import (
+    TRACK_GEOMETRY_FACTORY,
+    TrackGeometryDataset,
+)
 from OTAnalytics.plugin_datastore.track_geometry_store.pygeos_store import (
     PygeosTrackGeometryDataset,
 )
