+++ tests/OTAnalytics/plugin_datastore/test_track_store.py
@@ -11,13 +11,12 @@ from OTAnalytics.domain.geometry import (
     RelativeOffsetCoordinate,
     calculate_direction_vector,
 )
-from OTAnalytics.domain.track import (
+from OTAnalytics.domain.track import Track, TrackId
+from OTAnalytics.domain.track_dataset import (
     TRACK_GEOMETRY_FACTORY,
-    Track,
+    TrackDataset,
     TrackGeometryDataset,
-    TrackId,
 )
-from OTAnalytics.domain.track_dataset import TrackDataset
 from OTAnalytics.domain.types import EventType
 from OTAnalytics.plugin_datastore.python_track_store import (
     PythonTrack,
